--- testing/fleetserver/openapi/model_ack_response.go
@@ -1,56 +1,0 @@
-/*
- * fleet-server API
- *
- * The fleet-server API that is used by agents when enrolled with fleet.  Note that the current implementations in the fleet-server and elastic-agent may have some difference specifically when it comes to some objects. This is most notable when comparing the `Action` implementations. Fleet-server uses a general template for all actions and the elastic-agent will have more specific representations.  The implementation of fleet-server by default also includes a connection count limiter, as well as limiters for request body sizes. If an agent attempts to make request but there are no remaining connections, the attempt will be blocked and the agent will get an error. If an agent tries to send a body that is too large the fleet-server will respond with a 400 status code. 
- *
- * API version: 0.0.1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package openapi
-
-// AckResponse - Response to processing acknowledgement events.
-type AckResponse struct {
-
-	// The action result. Will have the value \"acks\".
-	Action string `json:"action"`
-
-	// A flag to indicate if one or more errors occured when proccessing events.
-	Errors bool `json:"errors"`
-
-	// The in-order list of results from processing events.
-	Items []AckResponseItem `json:"items"`
-}
-
-// AssertAckResponseRequired checks if the required fields are not zero-ed
-func AssertAckResponseRequired(obj AckResponse) error {
-	elements := map[string]interface{}{
-		"action": obj.Action,
-		"errors": obj.Errors,
-		"items": obj.Items,
-	}
-	for name, el := range elements {
-		if isZero := IsZeroValue(el); isZero {
-			return &RequiredError{Field: name}
-		}
-	}
-
-	for _, el := range obj.Items {
-		if err := AssertAckResponseItemRequired(el); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// AssertRecurseAckResponseRequired recursively checks if required fields are not zero-ed in a nested slice.
-// Accepts only nested slice of AckResponse (e.g. [][]AckResponse), otherwise ErrTypeAssertionError is thrown.
-func AssertRecurseAckResponseRequired(objSlice interface{}) error {
-	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
-		aAckResponse, ok := obj.(AckResponse)
-		if !ok {
-			return ErrTypeAssertionError
-		}
-		return AssertAckResponseRequired(aAckResponse)
-	})
-}
