--- testing/fleetserver/openapi/model_ack_response_item.go
@@ -1,46 +1,0 @@
-/*
- * fleet-server API
- *
- * The fleet-server API that is used by agents when enrolled with fleet.  Note that the current implementations in the fleet-server and elastic-agent may have some difference specifically when it comes to some objects. This is most notable when comparing the `Action` implementations. Fleet-server uses a general template for all actions and the elastic-agent will have more specific representations.  The implementation of fleet-server by default also includes a connection count limiter, as well as limiters for request body sizes. If an agent attempts to make request but there are no remaining connections, the attempt will be blocked and the agent will get an error. If an agent tries to send a body that is too large the fleet-server will respond with a 400 status code. 
- *
- * API version: 0.0.1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package openapi
-
-// AckResponseItem - The results of processing an acknowledgement event.
-type AckResponseItem struct {
-
-	// An HTTP status code that indicates if the event was processed successfully or not.
-	Status int32 `json:"status"`
-
-	// HTTP status text.
-	Message string `json:"message,omitempty"`
-}
-
-// AssertAckResponseItemRequired checks if the required fields are not zero-ed
-func AssertAckResponseItemRequired(obj AckResponseItem) error {
-	elements := map[string]interface{}{
-		"status": obj.Status,
-	}
-	for name, el := range elements {
-		if isZero := IsZeroValue(el); isZero {
-			return &RequiredError{Field: name}
-		}
-	}
-
-	return nil
-}
-
-// AssertRecurseAckResponseItemRequired recursively checks if required fields are not zero-ed in a nested slice.
-// Accepts only nested slice of AckResponseItem (e.g. [][]AckResponseItem), otherwise ErrTypeAssertionError is thrown.
-func AssertRecurseAckResponseItemRequired(objSlice interface{}) error {
-	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
-		aAckResponseItem, ok := obj.(AckResponseItem)
-		if !ok {
-			return ErrTypeAssertionError
-		}
-		return AssertAckResponseItemRequired(aAckResponseItem)
-	})
-}
