--- testing/fleetserver/openapi/model_action_signature.go
@@ -1,47 +1,0 @@
-/*
- * fleet-server API
- *
- * The fleet-server API that is used by agents when enrolled with fleet.  Note that the current implementations in the fleet-server and elastic-agent may have some difference specifically when it comes to some objects. This is most notable when comparing the `Action` implementations. Fleet-server uses a general template for all actions and the elastic-agent will have more specific representations.  The implementation of fleet-server by default also includes a connection count limiter, as well as limiters for request body sizes. If an agent attempts to make request but there are no remaining connections, the attempt will be blocked and the agent will get an error. If an agent tries to send a body that is too large the fleet-server will respond with a 400 status code. 
- *
- * API version: 0.0.1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package openapi
-
-// ActionSignature - Optional action signing data.
-type ActionSignature struct {
-
-	// The base64 encoded, UTF-8 JSON serialized action bytes that are signed.
-	Data string `json:"data" yaml:"data"`
-
-	// The base64 encoded signature.
-	Signature string `json:"signature" yaml:"signature"`
-}
-
-// AssertActionSignatureRequired checks if the required fields are not zero-ed
-func AssertActionSignatureRequired(obj ActionSignature) error {
-	elements := map[string]interface{}{
-		"data": obj.Data,
-		"signature": obj.Signature,
-	}
-	for name, el := range elements {
-		if isZero := IsZeroValue(el); isZero {
-			return &RequiredError{Field: name}
-		}
-	}
-
-	return nil
-}
-
-// AssertRecurseActionSignatureRequired recursively checks if required fields are not zero-ed in a nested slice.
-// Accepts only nested slice of ActionSignature (e.g. [][]ActionSignature), otherwise ErrTypeAssertionError is thrown.
-func AssertRecurseActionSignatureRequired(objSlice interface{}) error {
-	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
-		aActionSignature, ok := obj.(ActionSignature)
-		if !ok {
-			return ErrTypeAssertionError
-		}
-		return AssertActionSignatureRequired(aActionSignature)
-	})
-}
