--- testing/fleetserver/openapi/model_checkin_response.go
@@ -1,53 +1,0 @@
-/*
- * fleet-server API
- *
- * The fleet-server API that is used by agents when enrolled with fleet.  Note that the current implementations in the fleet-server and elastic-agent may have some difference specifically when it comes to some objects. This is most notable when comparing the `Action` implementations. Fleet-server uses a general template for all actions and the elastic-agent will have more specific representations.  The implementation of fleet-server by default also includes a connection count limiter, as well as limiters for request body sizes. If an agent attempts to make request but there are no remaining connections, the attempt will be blocked and the agent will get an error. If an agent tries to send a body that is too large the fleet-server will respond with a 400 status code. 
- *
- * API version: 0.0.1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package openapi
-
-type CheckinResponse struct {
-
-	// The acknowlegment token used to indicate action delivery.
-	AckToken string `json:"ack_token,omitempty"`
-
-	// The action result. Set to \"checkin\".
-	Action string `json:"action"`
-
-	// A list of actions that the agent must execute.
-	Actions []Action `json:"actions,omitempty"`
-}
-
-// AssertCheckinResponseRequired checks if the required fields are not zero-ed
-func AssertCheckinResponseRequired(obj CheckinResponse) error {
-	elements := map[string]interface{}{
-		"action": obj.Action,
-	}
-	for name, el := range elements {
-		if isZero := IsZeroValue(el); isZero {
-			return &RequiredError{Field: name}
-		}
-	}
-
-	for _, el := range obj.Actions {
-		if err := AssertActionRequired(el); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// AssertRecurseCheckinResponseRequired recursively checks if required fields are not zero-ed in a nested slice.
-// Accepts only nested slice of CheckinResponse (e.g. [][]CheckinResponse), otherwise ErrTypeAssertionError is thrown.
-func AssertRecurseCheckinResponseRequired(objSlice interface{}) error {
-	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
-		aCheckinResponse, ok := obj.(CheckinResponse)
-		if !ok {
-			return ErrTypeAssertionError
-		}
-		return AssertCheckinResponseRequired(aCheckinResponse)
-	})
-}
