--- testing/fleetserver/openapi/model_enroll_request.go
@@ -1,53 +1,0 @@
-/*
- * fleet-server API
- *
- * The fleet-server API that is used by agents when enrolled with fleet.  Note that the current implementations in the fleet-server and elastic-agent may have some difference specifically when it comes to some objects. This is most notable when comparing the `Action` implementations. Fleet-server uses a general template for all actions and the elastic-agent will have more specific representations.  The implementation of fleet-server by default also includes a connection count limiter, as well as limiters for request body sizes. If an agent attempts to make request but there are no remaining connections, the attempt will be blocked and the agent will get an error. If an agent tries to send a body that is too large the fleet-server will respond with a 400 status code. 
- *
- * API version: 0.0.1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package openapi
-
-// EnrollRequest - A request to enroll a new agent into fleet.
-type EnrollRequest struct {
-
-	// The enrollment type of the agent. The agent only supports the PERMANENT value 
-	Type string `json:"type"`
-
-	// The shared ID of the agent. To support pre-existing installs. NOT YET IMPLEMENTED. 
-	SharedId string `json:"shared_id"`
-
-	Metadata EnrollMetadata `json:"metadata"`
-}
-
-// AssertEnrollRequestRequired checks if the required fields are not zero-ed
-func AssertEnrollRequestRequired(obj EnrollRequest) error {
-	elements := map[string]interface{}{
-		"type": obj.Type,
-		"shared_id": obj.SharedId,
-		"metadata": obj.Metadata,
-	}
-	for name, el := range elements {
-		if isZero := IsZeroValue(el); isZero {
-			return &RequiredError{Field: name}
-		}
-	}
-
-	if err := AssertEnrollMetadataRequired(obj.Metadata); err != nil {
-		return err
-	}
-	return nil
-}
-
-// AssertRecurseEnrollRequestRequired recursively checks if required fields are not zero-ed in a nested slice.
-// Accepts only nested slice of EnrollRequest (e.g. [][]EnrollRequest), otherwise ErrTypeAssertionError is thrown.
-func AssertRecurseEnrollRequestRequired(objSlice interface{}) error {
-	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
-		aEnrollRequest, ok := obj.(EnrollRequest)
-		if !ok {
-			return ErrTypeAssertionError
-		}
-		return AssertEnrollRequestRequired(aEnrollRequest)
-	})
-}
