--- testing/fleetserver/openapi/model_status_response.go
@@ -1,52 +1,0 @@
-/*
- * fleet-server API
- *
- * The fleet-server API that is used by agents when enrolled with fleet.  Note that the current implementations in the fleet-server and elastic-agent may have some difference specifically when it comes to some objects. This is most notable when comparing the `Action` implementations. Fleet-server uses a general template for all actions and the elastic-agent will have more specific representations.  The implementation of fleet-server by default also includes a connection count limiter, as well as limiters for request body sizes. If an agent attempts to make request but there are no remaining connections, the attempt will be blocked and the agent will get an error. If an agent tries to send a body that is too large the fleet-server will respond with a 400 status code. 
- *
- * API version: 0.0.1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package openapi
-
-// StatusResponse - Status response information.
-type StatusResponse struct {
-
-	// Service name.
-	Name string `json:"name"`
-
-	// A Unit state that fleet-server may report. Unit state is defined in the elastic-agent-client specification. 
-	Status string `json:"status"`
-
-	Version StatusResponseVersion `json:"version,omitempty"`
-}
-
-// AssertStatusResponseRequired checks if the required fields are not zero-ed
-func AssertStatusResponseRequired(obj StatusResponse) error {
-	elements := map[string]interface{}{
-		"name": obj.Name,
-		"status": obj.Status,
-	}
-	for name, el := range elements {
-		if isZero := IsZeroValue(el); isZero {
-			return &RequiredError{Field: name}
-		}
-	}
-
-	if err := AssertStatusResponseVersionRequired(obj.Version); err != nil {
-		return err
-	}
-	return nil
-}
-
-// AssertRecurseStatusResponseRequired recursively checks if required fields are not zero-ed in a nested slice.
-// Accepts only nested slice of StatusResponse (e.g. [][]StatusResponse), otherwise ErrTypeAssertionError is thrown.
-func AssertRecurseStatusResponseRequired(objSlice interface{}) error {
-	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
-		aStatusResponse, ok := obj.(StatusResponse)
-		if !ok {
-			return ErrTypeAssertionError
-		}
-		return AssertStatusResponseRequired(aStatusResponse)
-	})
-}
