--- testing/fleetserver/openapi/model_upload_begin_request.go
@@ -1,53 +1,0 @@
-/*
- * fleet-server API
- *
- * The fleet-server API that is used by agents when enrolled with fleet.  Note that the current implementations in the fleet-server and elastic-agent may have some difference specifically when it comes to some objects. This is most notable when comparing the `Action` implementations. Fleet-server uses a general template for all actions and the elastic-agent will have more specific representations.  The implementation of fleet-server by default also includes a connection count limiter, as well as limiters for request body sizes. If an agent attempts to make request but there are no remaining connections, the attempt will be blocked and the agent will get an error. If an agent tries to send a body that is too large the fleet-server will respond with a 400 status code. 
- *
- * API version: 0.0.1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package openapi
-
-type UploadBeginRequest struct {
-
-	File UploadBeginRequestFile `json:"file"`
-
-	// ID of the action that requested this file
-	ActionId string `json:"action_id"`
-
-	// Identifier of the agent uploading. Matches the ID usually found in agent.id
-	AgentId string `json:"agent_id"`
-
-	// The source integration sending this file
-	Src string `json:"src"`
-}
-
-// AssertUploadBeginRequestRequired checks if the required fields are not zero-ed
-func AssertUploadBeginRequestRequired(obj UploadBeginRequest) error {
-	elements := map[string]interface{}{
-		"file": obj.File,
-		"action_id": obj.ActionId,
-		"agent_id": obj.AgentId,
-		"src": obj.Src,
-	}
-	for name, el := range elements {
-		if isZero := IsZeroValue(el); isZero {
-			return &RequiredError{Field: name}
-		}
-	}
-
-	return nil
-}
-
-// AssertRecurseUploadBeginRequestRequired recursively checks if required fields are not zero-ed in a nested slice.
-// Accepts only nested slice of UploadBeginRequest (e.g. [][]UploadBeginRequest), otherwise ErrTypeAssertionError is thrown.
-func AssertRecurseUploadBeginRequestRequired(objSlice interface{}) error {
-	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
-		aUploadBeginRequest, ok := obj.(UploadBeginRequest)
-		if !ok {
-			return ErrTypeAssertionError
-		}
-		return AssertUploadBeginRequestRequired(aUploadBeginRequest)
-	})
-}
