--- testing/fleetserver/openapi/model_upload_begin_request_file.go
@@ -1,58 +1,0 @@
-/*
- * fleet-server API
- *
- * The fleet-server API that is used by agents when enrolled with fleet.  Note that the current implementations in the fleet-server and elastic-agent may have some difference specifically when it comes to some objects. This is most notable when comparing the `Action` implementations. Fleet-server uses a general template for all actions and the elastic-agent will have more specific representations.  The implementation of fleet-server by default also includes a connection count limiter, as well as limiters for request body sizes. If an agent attempts to make request but there are no remaining connections, the attempt will be blocked and the agent will get an error. If an agent tries to send a body that is too large the fleet-server will respond with a 400 status code. 
- *
- * API version: 0.0.1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package openapi
-
-type UploadBeginRequestFile struct {
-
-	// The algorithm used to compress the file. Valid values: br,gzip,deflate,none
-	Compression string `json:"Compression,omitempty"`
-
-	Hash Hash `json:"hash,omitempty"`
-
-	// Name of the file including the extension, without the directory
-	Name string `json:"name"`
-
-	// MIME type of the file
-	MimeType string `json:"mime_type"`
-
-	// Size of the file contents, in bytes
-	Size int64 `json:"size"`
-}
-
-// AssertUploadBeginRequestFileRequired checks if the required fields are not zero-ed
-func AssertUploadBeginRequestFileRequired(obj UploadBeginRequestFile) error {
-	elements := map[string]interface{}{
-		"name": obj.Name,
-		"mime_type": obj.MimeType,
-		"size": obj.Size,
-	}
-	for name, el := range elements {
-		if isZero := IsZeroValue(el); isZero {
-			return &RequiredError{Field: name}
-		}
-	}
-
-	if err := AssertHashRequired(obj.Hash); err != nil {
-		return err
-	}
-	return nil
-}
-
-// AssertRecurseUploadBeginRequestFileRequired recursively checks if required fields are not zero-ed in a nested slice.
-// Accepts only nested slice of UploadBeginRequestFile (e.g. [][]UploadBeginRequestFile), otherwise ErrTypeAssertionError is thrown.
-func AssertRecurseUploadBeginRequestFileRequired(objSlice interface{}) error {
-	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
-		aUploadBeginRequestFile, ok := obj.(UploadBeginRequestFile)
-		if !ok {
-			return ErrTypeAssertionError
-		}
-		return AssertUploadBeginRequestFileRequired(aUploadBeginRequestFile)
-	})
-}
