--- testing/fleetserver/openapi/model_upload_begin_response.go
@@ -1,47 +1,0 @@
-/*
- * fleet-server API
- *
- * The fleet-server API that is used by agents when enrolled with fleet.  Note that the current implementations in the fleet-server and elastic-agent may have some difference specifically when it comes to some objects. This is most notable when comparing the `Action` implementations. Fleet-server uses a general template for all actions and the elastic-agent will have more specific representations.  The implementation of fleet-server by default also includes a connection count limiter, as well as limiters for request body sizes. If an agent attempts to make request but there are no remaining connections, the attempt will be blocked and the agent will get an error. If an agent tries to send a body that is too large the fleet-server will respond with a 400 status code. 
- *
- * API version: 0.0.1
- * Generated by: OpenAPI Generator (https://openapi-generator.tech)
- */
-
-package openapi
-
-// UploadBeginResponse - Response to initiating a file upload
-type UploadBeginResponse struct {
-
-	// A unique identifier for the ensuing upload operation
-	UploadId string `json:"upload_id"`
-
-	// The required size (in bytes) that the file must be segmented into for each chunk
-	ChunkSize int64 `json:"chunk_size"`
-}
-
-// AssertUploadBeginResponseRequired checks if the required fields are not zero-ed
-func AssertUploadBeginResponseRequired(obj UploadBeginResponse) error {
-	elements := map[string]interface{}{
-		"upload_id": obj.UploadId,
-		"chunk_size": obj.ChunkSize,
-	}
-	for name, el := range elements {
-		if isZero := IsZeroValue(el); isZero {
-			return &RequiredError{Field: name}
-		}
-	}
-
-	return nil
-}
-
-// AssertRecurseUploadBeginResponseRequired recursively checks if required fields are not zero-ed in a nested slice.
-// Accepts only nested slice of UploadBeginResponse (e.g. [][]UploadBeginResponse), otherwise ErrTypeAssertionError is thrown.
-func AssertRecurseUploadBeginResponseRequired(objSlice interface{}) error {
-	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
-		aUploadBeginResponse, ok := obj.(UploadBeginResponse)
-		if !ok {
-			return ErrTypeAssertionError
-		}
-		return AssertUploadBeginResponseRequired(aUploadBeginResponse)
-	})
-}
